DF-6051 warning fixes

+++ Projects/UnitTests/Classes/Tests/ComponentsTest.cpp
@@ -59,7 +59,7 @@ void SingleComponentSystem::AddEntity(Entity * entity)
     
     for(int32 id = 0; id < Component::COMPONENT_COUNT; ++id)
     {
-        uint32 flags = 1 << id;
+        uint64 flags = 1LL << id;
         if((flags & GetRequiredComponents()) == flags)
         {
             uint32 componentsCount = entity->GetComponentCount(id);
@@ -76,7 +76,7 @@ void SingleComponentSystem::RemoveEntity(Entity * entity)
 {
     for(int32 id = 0; id < Component::COMPONENT_COUNT; ++id)
     {
-        uint32 flags = 1 << id;
+        uint64 flags = 1LL << id;
         if((flags & GetRequiredComponents()) == flags)
         {
             uint32 componentsCount = entity->GetComponentCount(id);
@@ -123,7 +123,7 @@ void MultiComponentSystem::AddEntity(Entity * entity)
     
     for(int32 id = 0; id < Component::COMPONENT_COUNT; ++id)
     {
-        uint32 flags = 1 << id;
+        uint64 flags = 1LL << id;
         if((flags & GetRequiredComponents()) == flags)
         {
             uint32 componentsCount = entity->GetComponentCount(id);
@@ -140,7 +140,7 @@ void MultiComponentSystem::RemoveEntity(Entity * entity)
 {
     for(int32 id = 0; id < Component::COMPONENT_COUNT; ++id)
     {
-        uint32 flags = 1 << id;
+        uint64 flags = 1LL << id;
         if((flags & GetRequiredComponents()) == flags)
         {
             uint32 componentsCount = entity->GetComponentCount(id);
